Outsource mongodb actions
