+++ config/config.go
@@ -1,0 +1,44 @@
+package config
+
+// functions for accessing to mongodb
+import (
+	"context"
+	"fmt"
+	"log"
+	"os"
+
+	"github.com/joho/godotenv"
+	"go.mongodb.org/mongo-driver/mongo"
+	"go.mongodb.org/mongo-driver/mongo/options"
+	"go.mongodb.org/mongo-driver/mongo/readpref"
+)
+
+var DB *mongo.Database
+
+func Connect() *mongo.Database {
+	err := godotenv.Load(".env")
+	if err != nil {
+		log.Fatalf("Error loading .env file: %s", err)
+	}
+
+	MONGODB_URI := os.Getenv("MONGODB_URI")
+
+	clientOptions := options.Client().ApplyURI(MONGODB_URI)
+	client, err := mongo.Connect(context.Background(), clientOptions)
+	if err != nil {
+		log.Fatal(err)
+	}
+
+	err = client.Ping(context.TODO(), readpref.Primary())
+	if err != nil {
+		log.Fatal(err)
+	}
+
+	fmt.Println("Connected to MongoDB!")
+	return client.Database("sample_mflix")
+}
+
+// GetCollection returns a reference to a MongoDB collection
+func GetCollection(collectionName string) *mongo.Collection {
+	return DB.Collection(collectionName)
+}
