+++ go.mod
@@ -1,6 +1,6 @@
 module example/rest-api-demo
 
-go 1.19
+go 1.21
 
 require (
 	github.com/gin-gonic/gin v1.10.0
