+++ main.go
@@ -3,68 +3,42 @@ package main
 import (
 	"context"
 	"fmt"
-	"log"
 	"net/http"
-	"os"
 	"strconv"
 
 	"github.com/gin-gonic/gin"
-	"github.com/joho/godotenv"
 	"go.mongodb.org/mongo-driver/bson"
-	"go.mongodb.org/mongo-driver/bson/primitive"
-	"go.mongodb.org/mongo-driver/mongo"
 	"go.mongodb.org/mongo-driver/mongo/options"
-	"go.mongodb.org/mongo-driver/mongo/readpref"
+
+	"example/rest-api-demo/config"
 )
 
-var db *mongo.Database
+// var db *mongo.Database
 
 func main() {
-	db = Connect()
+	config.Connect()
 
 	server := gin.Default()
 	server.GET("/movies", GetMovies)
-	server.GET("/comments", GetComments)
-	server.GET("/comments/:id", GetCommentByID)
-	server.POST("/comments", CreateComment)
-	server.PUT("/comments/:id", UpdateCommentByID)
-	server.DELETE("/comments/:id", DeleteCommentByID)
+	// server.GET("/comments", GetComments)
+	// server.GET("/comments/:id", GetCommentByID)
+	// server.POST("/comments", CreateComment)
+	// server.PUT("/comments/:id", UpdateCommentByID)
+	// server.DELETE("/comments/:id", DeleteCommentByID)
 
-	server.GET("/movies/:id", GetMovieByID)
+	// server.GET("/movies/:id", GetMovieByID)
 	server.Run(":8080")
-}
-
-func Connect() *mongo.Database {
-	err := godotenv.Load(".env")
-	if err != nil {
-		log.Fatalf("Error loading .env file: %s", err)
-	}
-
-	MONGODB_URI := os.Getenv("MONGODB_URI")
-
-	clientOptions := options.Client().ApplyURI(MONGODB_URI)
-	client, err := mongo.Connect(context.Background(), clientOptions)
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	err = client.Ping(context.TODO(), readpref.Primary())
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	fmt.Println("Connected to MongoDB!")
-	return client.Database("sample_mflix")
 }
 
 func GetMovies(c *gin.Context) {
 	pageStr := c.Query("page")
 	sizeStr := c.Query("size")
 	page, err := strconv.Atoi(pageStr)
-	size, err := strconv.Atoi(sizeStr)
 	if err != nil || page < 1 {
 		page = 1
 	}
+
+	size, err := strconv.Atoi(sizeStr)
 
 	if err != nil || size < 1 {
 		size = 10
@@ -73,7 +47,7 @@ func GetMovies(c *gin.Context) {
 	limit := int64(10)
 	skip := int64((page - 1) * 10)
 
-	collection := db.Collection("movies")
+	collection := config.GetCollection("movies")
 	findOptions := options.Find()
 	findOptions.SetLimit(limit)
 	findOptions.SetSkip(skip)
@@ -94,6 +68,7 @@ func GetMovies(c *gin.Context) {
 	c.JSON(http.StatusOK, movies)
 }
 
+/*
 func GetComments(c *gin.Context) {
 	pageStr := c.Query("page")
 	sizeStr := c.Query("size")
@@ -246,3 +221,4 @@ func CreateComment(c *gin.Context) {
 
 	c.JSON(http.StatusCreated, gin.H{"message": "Comment created successfully", "id": result.InsertedID})
 }
+*/
